requirements.txt
pyrogram==2.0.106
tgcrypto==1.2.5
aiohttp==3.9.1
python-dotenv==1.0.0
pytz==2023.3